static/requirements.txt
djangorestframework
djangorestframework-simplejwt
drf-yasg
psycopg2-binary
python-dotenv
Pillow
Django>=3.2,<4.0
django-cors-headers
django-filter